KP-10352 Move general use CLI tests to conftest
This allows reusing them later for further testing
